Module Posix_time2.Tm

POSIX broken-down time structure.

type t = private {

tm_sec : int;

tm_min : int;

tm_hour : int;

tm_mday : int;

tm_mon : int;

tm_year : int;

tm_wday : int;

tm_yday : int;

tm_isdst : int;

}

Broken-down time with fields for seconds 0-60, minutes 0-59, hours 0-23, day of month 1-31, month 0-11, years since 1900, day of week 0-6 (Sunday = 0), day of year 0-365, and daylight saving flag (positive if DST, 0 if not, negative if unknown).

val create : int -> int -> int -> int -> int -> int -> int -> int -> int -> t option

create sec min hour mday mon year wday yday isdst creates a new broken-down time value. Returns None if arguments are invalid.

val compare : t -> t -> int

compare t1 t2 compares two time values considering year, month, day, hour, minute, second in that order.